fix: warnings from gcc with cpu backend

+++ src/runko/emf/tile.c++
@@ -177,16 +177,23 @@ Tile<D>::Tile(
   // Tile size (can be used as yee_lattice_ is already constructed from N{x,y,z}Mesh.
   const auto [Lx, Ly, Lz] = yee_lattice_.extents_wout_halo();
 
-  this->set_tile_mins({ xmin + i * Lx, ymin + j * Ly, zmin + k * Lz });
+  this->set_tile_mins(
+    { xmin + static_cast<double>(i * Lx),
+      ymin + static_cast<double>(j * Ly),
+      zmin + static_cast<double>(k * Lz) });
   this->set_tile_maxs(
-    { xmin + (i + 1uz) * Lx, ymin + (j + 1uz) * Ly, zmin + (k + 1uz) * Lz });
+    { xmin + static_cast<double>((i + 1uz) * Lx),
+      ymin + static_cast<double>((j + 1uz) * Ly),
+      zmin + static_cast<double>((k + 1uz) * Lz) });
 
 
   this->global_coordinate_mins_ = { xmin, ymin, zmin };
 
-  this->global_coordinate_maxs_ = { global_coordinate_mins_[0] + Nx * Lx,
-                                    global_coordinate_mins_[1] + Ny * Ly,
-                                    global_coordinate_mins_[2] + Nz * Lz };
+  this->global_coordinate_maxs_ = {
+    global_coordinate_mins_[0] + static_cast<double>(Nx * Lx),
+    global_coordinate_mins_[1] + static_cast<double>(Ny * Ly),
+    global_coordinate_mins_[2] + static_cast<double>(Nz * Lz)
+  };
 }
 
 template<std::size_t D>
